refactor: :recycle: detach player classes from game.py
